refactor(zkvm): split into multiple files
