+++ crates/zkvm/zkvm/src/proof.rs
@@ -1,0 +1,60 @@
+use arbitrary::Arbitrary;
+use borsh::{BorshDeserialize, BorshSerialize};
+use serde::{Deserialize, Serialize};
+
+use crate::VerificationKey;
+
+/// Validity proof generated by the `ZKVMHost`
+#[derive(
+    Debug, Clone, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq, Arbitrary,
+)]
+pub struct Proof(Vec<u8>);
+
+impl Proof {
+    pub fn new(data: Vec<u8>) -> Self {
+        Self(data)
+    }
+
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+}
+
+impl From<Proof> for Vec<u8> {
+    fn from(value: Proof) -> Self {
+        value.0
+    }
+}
+
+impl From<&Proof> for Vec<u8> {
+    fn from(value: &Proof) -> Self {
+        value.0.clone()
+    }
+}
+
+/// An input to the aggregation program.
+///
+/// Consists of a proof and a verification key.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct AggregationInput {
+    proof: Proof,
+    vk: VerificationKey,
+}
+
+impl AggregationInput {
+    pub fn new(proof: Proof, vk: VerificationKey) -> Self {
+        Self { proof, vk }
+    }
+
+    pub fn proof(&self) -> &Proof {
+        &self.proof
+    }
+
+    pub fn vk(&self) -> &VerificationKey {
+        &self.vk
+    }
+}
