+++ crates/zkvm/zkvm/src/verifer.rs
@@ -1,0 +1,52 @@
+use arbitrary::Arbitrary;
+use borsh::{BorshDeserialize, BorshSerialize};
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
+
+use crate::Proof;
+
+/// A trait implemented by a verifier to decode and verify the proof generated by the prover
+/// ("host").
+pub trait ZKVMVerifier {
+    /// Verifies the proof generated by the prover against the `program_id`.
+    fn verify(verification_key: &VerificationKey, proof: &Proof) -> anyhow::Result<()>;
+
+    /// Verifies the proof generated by the prover against the given `program_id` and
+    /// `public_params`.
+    fn verify_with_public_params<T: Serialize + DeserializeOwned>(
+        verification_key: &VerificationKey,
+        public_params: T,
+        proof: &Proof,
+    ) -> anyhow::Result<()>;
+
+    fn verify_groth16(
+        proof: &Proof,
+        verification_key: &[u8],
+        public_params_raw: &[u8],
+    ) -> anyhow::Result<()>;
+
+    /// Extracts the public output from the given proof using standard `serde`
+    /// serialization/deserialization.
+    fn extract_public_output<T: Serialize + DeserializeOwned>(proof: &Proof) -> anyhow::Result<T>;
+
+    /// Extracts the public output from the given proof assuming the data was serialized using
+    /// Borsh.
+    fn extract_borsh_public_output<T: BorshSerialize + BorshDeserialize>(
+        proof: &Proof,
+    ) -> anyhow::Result<T>;
+}
+
+/// Verification Key required to verify proof generated from `ZKVMHost`
+#[derive(
+    Debug, Clone, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq, Arbitrary,
+)]
+pub struct VerificationKey(pub Vec<u8>);
+
+impl VerificationKey {
+    pub fn new(data: Vec<u8>) -> Self {
+        Self(data)
+    }
+
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+}
